backend/requirements.txt
fastapi
uvicorn
whisper
google-generativeai
python-dotenv
pydantic
ffmpeg-python
